move graphql tests (#36475)
